+++ 2.Logic-Layer/PresupuestosServices/src/vo/vista/UserListJSON.java
@@ -1,7 +1,7 @@
 package vo.vista;
 
 public class UserListJSON {
-	private int iduser;
+	private long iduser;
 	private long document;
 	private String name;
 	private String username;
@@ -15,10 +15,10 @@ public class UserListJSON {
 	public void setEmail(String email) {
 		this.email = email;
 	}
-	public int getIduser() {
+	public long getIduser() {
 		return iduser;
 	}
-	public void setIduser(int iduser) {
+	public void setIduser(long iduser) {
 		this.iduser = iduser;
 	}
 	public long getDocument() {
