+++ 3.Persistence-Layer/spdb.sql
@@ -18,7 +18,7 @@ USE `SPDB` ;
 -- Table `SPDB`.`Area`
 -- -----------------------------------------------------
 CREATE TABLE IF NOT EXISTS `SPDB`.`Area` (
-  `idArea` INT NOT NULL AUTO_INCREMENT,
+  `idArea` BIGINT(64) NOT NULL AUTO_INCREMENT,
   `name` VARCHAR(45) NOT NULL,
   PRIMARY KEY (`idArea`))
 ENGINE = InnoDB;
@@ -28,12 +28,12 @@ ENGINE = InnoDB;
 -- Table `SPDB`.`User`
 -- -----------------------------------------------------
 CREATE TABLE IF NOT EXISTS `SPDB`.`User` (
-  `idUser` INT NOT NULL AUTO_INCREMENT,
-  `document` INT NOT NULL,
+  `idUser` BIGINT(64) NOT NULL AUTO_INCREMENT,
+  `document` BIGINT(64) NOT NULL,
   `name` VARCHAR(100) NOT NULL,
   `userName` VARCHAR(100) NOT NULL,
   `password` VARCHAR(45) NOT NULL,
-  `idArea` INT NOT NULL,
+  `idArea` BIGINT(64) NOT NULL,
   `email` VARCHAR(100) NULL,
   PRIMARY KEY (`idUser`),
   INDEX `fk_User_Area_idx` (`idArea` ASC),
@@ -49,7 +49,7 @@ ENGINE = InnoDB;
 -- Table `SPDB`.`Role`
 -- -----------------------------------------------------
 CREATE TABLE IF NOT EXISTS `SPDB`.`Role` (
-  `idRole` INT NOT NULL AUTO_INCREMENT,
+  `idRole` BIGINT(64) NOT NULL AUTO_INCREMENT,
   `name` VARCHAR(45) NOT NULL,
   `description` TEXT NULL,
   PRIMARY KEY (`idRole`))
@@ -60,9 +60,9 @@ ENGINE = InnoDB;
 -- Table `SPDB`.`User_Role`
 -- -----------------------------------------------------
 CREATE TABLE IF NOT EXISTS `SPDB`.`User_Role` (
-  `idUser_Role` INT NOT NULL AUTO_INCREMENT,
-  `idUser` INT NOT NULL,
-  `idRole` INT NOT NULL,
+  `idUser_Role` BIGINT(64) NOT NULL AUTO_INCREMENT,
+  `idUser` BIGINT(64) NOT NULL,
+  `idRole` BIGINT(64) NOT NULL,
   PRIMARY KEY (`idUser_Role`),
   INDEX `fk_User_Role_User1_idx` (`idUser` ASC),
   INDEX `fk_User_Role_Role1_idx` (`idRole` ASC),
@@ -83,7 +83,7 @@ ENGINE = InnoDB;
 -- Table `SPDB`.`Provider`
 -- -----------------------------------------------------
 CREATE TABLE IF NOT EXISTS `SPDB`.`Provider` (
-  `idProvider` INT NOT NULL AUTO_INCREMENT,
+  `idProvider` BIGINT(64) NOT NULL AUTO_INCREMENT,
   `NIT` VARCHAR(45) NOT NULL,
   `name` VARCHAR(45) NOT NULL,
   `description` TEXT NULL,
@@ -95,7 +95,7 @@ ENGINE = InnoDB;
 -- Table `SPDB`.`Country`
 -- -----------------------------------------------------
 CREATE TABLE IF NOT EXISTS `SPDB`.`Country` (
-  `idCountry` INT NOT NULL AUTO_INCREMENT,
+  `idCountry` BIGINT(64) NOT NULL AUTO_INCREMENT,
   `countryCode` VARCHAR(5) NOT NULL,
   `name` VARCHAR(45) NOT NULL,
   PRIMARY KEY (`idCountry`))
@@ -106,9 +106,9 @@ ENGINE = InnoDB;
 -- Table `SPDB`.`City`
 -- -----------------------------------------------------
 CREATE TABLE IF NOT EXISTS `SPDB`.`City` (
-  `idCity` INT NOT NULL AUTO_INCREMENT,
+  `idCity` BIGINT(64) NOT NULL AUTO_INCREMENT,
   `name` VARCHAR(45) NOT NULL,
-  `idCountry` INT NOT NULL,
+  `idCountry` BIGINT(64) NOT NULL,
   PRIMARY KEY (`idCity`),
   INDEX `fk_City_Country1_idx` (`idCountry` ASC),
   CONSTRAINT `fk_City_Country1`
@@ -123,7 +123,7 @@ ENGINE = InnoDB;
 -- Table `SPDB`.`Client`
 -- -----------------------------------------------------
 CREATE TABLE IF NOT EXISTS `SPDB`.`Client` (
-  `idClient` INT NOT NULL AUTO_INCREMENT,
+  `idClient` BIGINT(64) NOT NULL AUTO_INCREMENT,
   `NIT` VARCHAR(45) NULL,
   `name` VARCHAR(45) NOT NULL,
   `description` TEXT NULL,
@@ -135,11 +135,11 @@ ENGINE = InnoDB;
 -- Table `SPDB`.`Address`
 -- -----------------------------------------------------
 CREATE TABLE IF NOT EXISTS `SPDB`.`Address` (
-  `idAddress` INT NOT NULL AUTO_INCREMENT,
+  `idAddress` BIGINT(64) NOT NULL AUTO_INCREMENT,
   `address` VARCHAR(45) NOT NULL,
-  `idProvider` INT NULL,
-  `idCity` INT NOT NULL,
-  `idClient` INT NULL,
+  `idProvider` BIGINT(64) NULL,
+  `idCity` BIGINT(64) NOT NULL,
+  `idClient` BIGINT(64) NULL,
   PRIMARY KEY (`idAddress`),
   INDEX `fk_Address_Provider1_idx` (`idProvider` ASC),
   INDEX `fk_Address_City1_idx` (`idCity` ASC),
@@ -166,12 +166,12 @@ ENGINE = InnoDB;
 -- Table `SPDB`.`Contact`
 -- -----------------------------------------------------
 CREATE TABLE IF NOT EXISTS `SPDB`.`Contact` (
-  `idContact` INT NOT NULL AUTO_INCREMENT,
+  `idContact` BIGINT(64) NOT NULL AUTO_INCREMENT,
   `name` VARCHAR(100) NOT NULL,
   `email` VARCHAR(100) NOT NULL,
   `phoneNumber` VARCHAR(100) NOT NULL,
-  `idProvider` INT NULL,
-  `idClient` INT NULL,
+  `idProvider` BIGINT(64) NULL,
+  `idClient` BIGINT(64) NULL,
   PRIMARY KEY (`idContact`),
   INDEX `fk_Contact_Provider1_idx` (`idProvider` ASC),
   INDEX `fk_Contact_Client1_idx` (`idClient` ASC),
@@ -192,11 +192,11 @@ ENGINE = InnoDB;
 -- Table `SPDB`.`ProductService`
 -- -----------------------------------------------------
 CREATE TABLE IF NOT EXISTS `SPDB`.`ProductService` (
-  `idProductService` INT NOT NULL AUTO_INCREMENT,
+  `idProductService` BIGINT(64) NOT NULL AUTO_INCREMENT,
   `name` VARCHAR(45) NOT NULL,
   `description` TEXT NULL,
-  `price` DOUBLE NOT NULL,
-  `idProvider` INT NOT NULL,
+  `price` DECIMAL NOT NULL,
+  `idProvider` BIGINT(64) NOT NULL,
   PRIMARY KEY (`idProductService`),
   INDEX `fk_ProductService_Provider1_idx` (`idProvider` ASC),
   CONSTRAINT `fk_ProductService_Provider1`
@@ -211,10 +211,10 @@ ENGINE = InnoDB;
 -- Table `SPDB`.`Project`
 -- -----------------------------------------------------
 CREATE TABLE IF NOT EXISTS `SPDB`.`Project` (
-  `idProject` INT NOT NULL AUTO_INCREMENT,
+  `idProject` BIGINT(64) NOT NULL AUTO_INCREMENT,
   `name` VARCHAR(45) NOT NULL,
-  `idClient` INT NOT NULL,
-  `User_idUser` INT NOT NULL,
+  `idClient` BIGINT(64) NOT NULL,
+  `User_idUser` BIGINT(64) NOT NULL,
   PRIMARY KEY (`idProject`),
   INDEX `fk_Project_Client1_idx` (`idClient` ASC),
   INDEX `fk_Project_User1_idx` (`User_idUser` ASC),
@@ -235,15 +235,15 @@ ENGINE = InnoDB;
 -- Table `SPDB`.`Budget`
 -- -----------------------------------------------------
 CREATE TABLE IF NOT EXISTS `SPDB`.`Budget` (
-  `idBudget` INT NOT NULL AUTO_INCREMENT,
+  `idBudget` BIGINT(64) NOT NULL AUTO_INCREMENT,
   `observations` TEXT NULL,
   `date` DATE NOT NULL,
   `commercialConditions` TEXT NULL,
-  `bruteTotal` DOUBLE NOT NULL,
-  `IVA` DOUBLE NOT NULL,
+  `bruteTotal` DECIMAL NOT NULL,
+  `IVA` DECIMAL NOT NULL,
   `months` INT NOT NULL,
-  `activityTotal` DOUBLE NOT NULL,
-  `idProject` INT NOT NULL,
+  `activityTotal` DECIMAL NOT NULL,
+  `idProject` BIGINT(64) NOT NULL,
   PRIMARY KEY (`idBudget`),
   INDEX `fk_Budget_Project1_idx` (`idProject` ASC),
   CONSTRAINT `fk_Budget_Project1`
@@ -258,13 +258,13 @@ ENGINE = InnoDB;
 -- Table `SPDB`.`BudgetPS`
 -- -----------------------------------------------------
 CREATE TABLE IF NOT EXISTS `SPDB`.`BudgetPS` (
-  `idBudgetPS` INT NOT NULL AUTO_INCREMENT,
-  `margin` DOUBLE NOT NULL,
-  `amount` INT NOT NULL,
+  `idBudgetPS` BIGINT(64) NOT NULL AUTO_INCREMENT,
+  `margin` DECIMAL NOT NULL,
+  `amount` BIGINT(64) NOT NULL,
   `days` INT NOT NULL,
-  `unitValue` DOUBLE NOT NULL,
-  `idProductService` INT NOT NULL,
-  `idBudget` INT NOT NULL,
+  `unitValue` DECIMAL NOT NULL,
+  `idProductService` BIGINT(64) NOT NULL,
+  `idBudget` BIGINT(64) NOT NULL,
   PRIMARY KEY (`idBudgetPS`),
   INDEX `fk_BudgetPS_ProductService1_idx` (`idProductService` ASC),
   INDEX `fk_BudgetPS_Budget1_idx` (`idBudget` ASC),
@@ -285,11 +285,11 @@ ENGINE = InnoDB;
 -- Table `SPDB`.`Expenses`
 -- -----------------------------------------------------
 CREATE TABLE IF NOT EXISTS `SPDB`.`Expenses` (
-  `idExpenses` INT NOT NULL AUTO_INCREMENT,
+  `idExpenses` BIGINT(64) NOT NULL AUTO_INCREMENT,
   `name` VARCHAR(45) NOT NULL,
   `description` TEXT NULL,
-  `value` DOUBLE NOT NULL,
-  `idBudgetPS` INT NOT NULL,
+  `value` DECIMAL NOT NULL,
+  `idBudgetPS` BIGINT(64) NOT NULL,
   PRIMARY KEY (`idExpenses`),
   INDEX `fk_Expenses_BudgetPS1_idx` (`idBudgetPS` ASC),
   CONSTRAINT `fk_Expenses_BudgetPS1`
@@ -304,9 +304,9 @@ ENGINE = InnoDB;
 -- Table `SPDB`.`Value`
 -- -----------------------------------------------------
 CREATE TABLE IF NOT EXISTS `SPDB`.`Value` (
-  `idValue` INT NOT NULL AUTO_INCREMENT,
+  `idValue` BIGINT(64) NOT NULL AUTO_INCREMENT,
   `name` VARCHAR(45) NOT NULL,
-  `value` INT NOT NULL,
+  `value` DECIMAL NOT NULL,
   PRIMARY KEY (`idValue`))
 ENGINE = InnoDB;
